requirements.txt
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
iniconfig==2.3.0
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
pip==25.3
pluggy==1.6.0
pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pyyaml==6.0.3
scipy==1.16.3
six==1.17.0
tqdm==4.67.1
